move sun example to experimental folder
