+++ examples/experimental/sun/src/shadow-effect/shadow-effect.js
