+++ examples/experimental/sun/src/shadow-effect/shadow-module.js
